Add shared sidebar to all Track pages and simplify testing card
- Create TrackSidebar component for consistent navigation
- Add sidebar to category pages with active state highlighting
- Remove breadcrumbs from category pages
- Simplify RiscVTestingCard to just link to test monitor
- Rename "zkVM Integration" to "ZKVM Security"
